requirements.txt
Werkzeug==2.0.1
Flask==2.0.3
Flask-RESTful==0.3.9
numpy==1.19.1
Keras==2.3.1
gevent==20.9.0
greenlet==0.4.17
pillow==7.2.0
h5py==2.10.0
opencv-python-headless==4.2.0.34
flask_pymongo==2.3.0
dnspython==1.16.0
urllib3==1.25.10
prometheus-flask-exporter==0.18.6